Overhauled Code
I have majorly updated the code to finally fix my past mistakes. Instead of 'caching'  the terms by saving the images of the individual tiles, now all of the terms are saved in a text document and the tiles are never saved. I also removed the requirement for the Free Space image to be present in the folder with the code. It required a LOT of time and a LOT of rewriting but I think it's finally finished. Thanks to the new method of saving the terms it is just as fast as before but uses significantly less space to do so.